+++ spotbugs-tests/src/test/java/edu/umd/cs/findbugs/ba/NullnessAnnotationTest.java
@@ -41,6 +41,11 @@ class NullnessAnnotationTest {
                 Arguments.of(javax.annotation.Nonnull.class.getName(), NullnessAnnotation.NONNULL),
                 Arguments.of(javax.annotation.Nullable.class.getName(), NullnessAnnotation.NULLABLE),
 
+                Arguments.of("org.jspecify.annotations.Nullable", NullnessAnnotation.CHECK_FOR_NULL),
+                Arguments.of("org.jspecify.annotations.NonNull", NullnessAnnotation.NONNULL),
+                Arguments.of("org.jspecify.annotations.NullUnmarked", NullnessAnnotation.UNKNOWN_NULLNESS),
+                Arguments.of("org.jspecify.annotations.NullMarked", NullnessAnnotation.NONNULL),
+
                 Arguments.of("something different", null));
     }